+++ src/main/java/net/Gabou/projectatmosphere/modules/temperature/spike/SpikeData.java
@@ -1,6 +1,6 @@
 package net.Gabou.projectatmosphere.modules.temperature.spike;
 
-import net.Gabou.projectatmosphere.util.BiomeInstanceKey;
+import net.Gabou.projectatmosphere.util.RegionInstanceKey;
 
-public record SpikeData(BiomeInstanceKey biome, float[][] week, SpikeState state) {
+public record SpikeData(RegionInstanceKey regionId, float[][] week, SpikeState state) {
 }