+++ src/main/java/net/Gabou/projectatmosphere/modules/temperature/spike/SpikeStateStorage.java
@@ -1,6 +1,8 @@
 package net.Gabou.projectatmosphere.modules.temperature.spike;
 
 import com.google.gson.*;
+import net.Gabou.projectatmosphere.util.RegionInstanceKey;
+import net.Gabou.projectatmosphere.modules.region.RegionIdCodec;
 import net.Gabou.projectatmosphere.util.AtmosphereUtils;
 import net.Gabou.projectatmosphere.util.BiomeInstanceKey;
 import net.Gabou.projectatmosphere.util.StorageUtils;
@@ -24,19 +26,20 @@ public class SpikeStateStorage {
     public static void saveAll(ServerLevel world) {
             JsonObject root = new JsonObject();
 
-            for (Map.Entry<BiomeInstanceKey, SpikeState> entry : SpikeManager.getAllStates().entrySet()) {
+            for (Map.Entry<RegionInstanceKey, SpikeState> entry : SpikeManager.getAllStates().entrySet()) {
                 JsonObject data = new JsonObject();
                 SpikeState state = entry.getValue();
 
-                data.addProperty("biome", entry.getKey().biomeType().toString());
-                data.add("pos", AtmosphereUtils.serializeBlockPos(entry.getKey().samplePos()));
+                data.addProperty("rx", entry.getKey().regionX());
+                data.addProperty("rz", entry.getKey().regionZ());
+                data.addProperty("size", entry.getKey().regionSize());
 
                 data.addProperty("daysSinceLastSpike", state.daysSinceLastSpike);
                 data.addProperty("remainingSpikeDays", state.remainingSpikeDays);
                 data.addProperty("currentSpikeDay", state.currentSpikeDay);
                 data.addProperty("spikeMagnitude", state.spikeMagnitude);
 
-                root.add(entry.getKey().toString(), data);
+                root.add(entry.getKey().regionX() + "," + entry.getKey().regionZ() + "@" + entry.getKey().regionSize(), data);
             }
 
             try {
@@ -60,15 +63,15 @@ public class SpikeStateStorage {
 
             try (Reader r = Files.newBufferedReader(SAVE_PATH)) {
                 JsonObject root = GSON.fromJson(r, JsonObject.class);
-                Map<BiomeInstanceKey, SpikeState> loaded = new HashMap<>();
+                Map<RegionInstanceKey, SpikeState> loaded = new HashMap<>();
 
                 for (Map.Entry<String, JsonElement> entry : root.entrySet()) {
                     JsonObject obj = entry.getValue().getAsJsonObject();
 
-                    ResourceLocation biome = ResourceLocation.parse(obj.get("biome").getAsString());
-                    BlockPos pos = AtmosphereUtils.deserializeBlockPos(obj.get("pos").getAsJsonObject());
-
-                    BiomeInstanceKey key = new BiomeInstanceKey(biome, pos);
+                    int rx = obj.get("rx").getAsInt();
+                    int rz = obj.get("rz").getAsInt();
+                    int size = obj.get("size").getAsInt();
+                    RegionInstanceKey key = new RegionInstanceKey(rx, rz, size);
 
                     SpikeState state = new SpikeState();
                     state.daysSinceLastSpike = obj.get("daysSinceLastSpike").getAsInt();